NgRx Metareducers
A metareducer is processed before the others reducers are invoked - when an action is triggered, ngrx triggers the configured metareducers before actually processing the action and the reducers involved. When all metareducers are finished, only then do the actions proceed (p.s. metareducers are invoked in a specific order all the time, depending on which ones come first in the metaReducers array)